atoi & strdup

+++ libc/string/strlen.c
@@ -1,8 +1,8 @@
 #include <string.h>
 
-size_t strlen(const char* string)
+int strlen(const char* string)
 {
-    size_t result = 0;
+    int result = 0;
     while ( string[result] )
         result++;
     return result;
